+++ CHANGELOG.md
@@ -5,8 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## [6.1.41] - 2025-01-16
 ### Changed
 - Hide Connections (messaging) from Sections and Directory [#4622](https://github.com/rokwire/illinois-app/issues/4622).
+- GroupDetailPanel UI updates (in progress) [#4483](https://github.com/rokwire/illinois-app/issues/4483).
+### Added
+- Show post replies count [#4618](https://github.com/rokwire/illinois-app/issues/4618).
+- Link previews for videos in group posts [#4581](https://github.com/rokwire/illinois-app/issues/4581)
 
 ## [6.1.40] - 2025-01-15
 ### Changed